+++ beacon-chain/core/state/state.go
@@ -50,7 +50,7 @@ func InitialBeaconState(
 	latestCrosslinks := make([]*pb.CrosslinkRecord, config.ShardCount)
 	for i := 0; i < len(latestCrosslinks); i++ {
 		latestCrosslinks[i] = &pb.CrosslinkRecord{
-			Slot:                 config.GenesisSlot,
+			Epoch:                config.GenesisEpoch,
 			ShardBlockRootHash32: config.ZeroHash[:],
 		}
 	}
