+++ beacon-chain/core/validators/BUILD.bazel
@@ -2,19 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "committees.go",
-        "validator.go",
-    ],
+    srcs = ["validator.go"],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/validators",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//beacon-chain/core/helpers:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/bitutil:go_default_library",
         "//shared/bytesutil:go_default_library",
         "//shared/hashutil:go_default_library",
-        "//shared/mathutil:go_default_library",
         "//shared/params:go_default_library",
         "//shared/sliceutil:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
@@ -23,10 +18,7 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "committees_test.go",
-        "validator_test.go",
-    ],
+    srcs = ["validator_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/core/helpers:go_default_library",
